// src/modules/permissions/enums/system-permission.enum.ts
enum SystemPermission {
  PreviewUserList = 100,
  PreviewUserProfile = 101,
  AddUser = 102,
  EditUserProfile = 103,
  DeactivateUser = 104,
  ActivateUser = 105,
  DeleteUser = 106,
  UnlockUser = 107,

  AddPermission = 200,
  DeletePermission = 201,
}

export { SystemPermission };